% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModFit.R
\name{ModFit}
\alias{ModFit}
\title{Model Estimation with either Group or Individual Penalties}
\usage{
ModFit(y, x, xGsz, lam1, lamG)
}
\arguments{
\item{y}{Response variables, it can be either a vector (univariate) or a
matrix (multivariate).}

\item{x}{Predictors, a matrix of dimensions N * p. There is a group structure
within x, and columns that belong to the same group should be adjacent to each other.}

\item{xGsz}{Size of each sub-group in x, it is a vector, the length of which is the
number of groups. To be more specific, if \eqn{xGsz=(g_1,..,g_m)}, then it means
the first \eqn{g_1} columns forms a group, the next \eqn{g_2} colums form another group, and so on.}

\item{lam1}{Individual penalty (a scalar), imposed to all coefficients.}

\item{lamG}{Group penalites (a vector), each entry is the group penalty imposed
on a sub-group of coefficients, corresponding to the sub-group in x.}
}
\value{
Returns an list with \item{res}{The fitted residuals. It is a matrix in the
case of multivariate regression. Otherwise it is just a vector.}
\item{Beta}{The fitted coefficients. It can be a matrix (multivariate case)
or a vector (univariate case)}
}
\description{
This function fits the regression model by regularized method to get the fitted
residuals and coefficients.
}
