% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneHot_Protein.R
\name{OneHot_Protein}
\alias{OneHot_Protein}
\title{MSA Data Preprocessing}
\usage{
OneHot_Protein(data)
}
\arguments{
\item{data}{The MSA data should be a N*m matrix, with N sequences and m positions.
Each amino acid is encoded by a number from 1 to 20, while 0 represents the gap.}
}
\value{
Returns an list with \item{data1hot}{The one hot encoded data. Note that the columns
representing gaps are removed, so are columns with only 0 entries. The standardized
data are reported in the end.}
\item{Gsz}{A vector of length m. The binary vectors belonging to the same position form
a group, and each entry of Gsz is the group size for the associated position. }
}
\description{
This function one hot encodes the original protein MSA sequence data.
}
