\name{SixProteinFamilies}

\alias{SixProteinFamilies}

\docType{data}

\title{
   Six Protein Families
}

\description{
 MSA data for six different protein families. 0 indicates the alignment gaps, while the other numbers represent amino acids.
}



\author{Yanming Li, Bin Nan, Ji Zhu}

\examples{
data(Beta.m)
}

\keyword{datasets}
