% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestLS.R
\name{TestLS}
\alias{TestLS}
\title{\eqn{\ell_2} Norm and Sup Norm Based Tests}
\usage{
TestLS(res1, res2)
}
\arguments{
\item{res1, res2}{Two data matrice with the same sample size. The goal is to test
if the covariance between them are 0 or not.}
}
\value{
Returns an list with \item{tstats_sup}{The Sup norm based test statistic.}
\item{pval_sup}{The corresponding p-value for the Sup norm based statistic. The
reference distribution is Gumbel.}
\item{tstats_chisq}{The \eqn{\ell_2} norm based test statistic.}
\item{pval_chisq}{The corresponding p-value for the \eqn{\ell_2} norm based statistic. The
reference distribution is weighted Chi-square.}
\item{cov_selfnorm}{The entry-wise self-normalized partial covariance matrix, based
on which, the \eqn{\ell_2} and Sup norm based statistics are constructed.}
\item{weights}{The estimated weights for the weighted Chi-square of reference.}
}
\description{
This function use both \eqn{\ell_2} and Sup norm based statistics
to infer if covariance between two random vectors are zero or not.
}
